server/requirements.txt
asgiref==3.11.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
imbalanced-learn==0.14.1
joblib==1.5.3
numpy==2.4.1
packaging==26.0
pandas==3.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sklearn-compat==0.1.5
sqlparse==0.5.5
threadpoolctl==3.6.0
whitenoise==6.11.0
